refactor: move ContextExt from context_adapter to context module
- Moved ContextExt trait from context_adapter.rs to context.rs
- Updated all imports to reference the new location
- Removed now-empty context_adapter.rs module
- Updated mod.rs to remove context_adapter module declaration
- All tests passing after the refactoring
