+++ crates/q_cli/src/cli/chat/commands/context/remove.rs
@@ -14,7 +14,7 @@ use eyre::{
 use fig_os_shim::Context;
 
 use crate::cli::chat::commands::CommandHandler;
-use crate::cli::chat::context_adapter::ContextExt;
+use crate::cli::chat::context::ContextExt;
 use crate::cli::chat::{
     ChatState,
     QueuedTool,
