+++ crates/q_cli/src/cli/chat/commands/context/show.rs
@@ -11,7 +11,7 @@ use eyre::Result;
 use fig_os_shim::Context;
 
 use crate::cli::chat::commands::CommandHandler;
-use crate::cli::chat::context_adapter::ContextExt;
+use crate::cli::chat::context::ContextExt;
 use crate::cli::chat::{
     ChatState,
     QueuedTool,
