+++ crates/q_cli/src/cli/chat/mod.rs
@@ -1,7 +1,6 @@
 mod command;
 mod commands;
 mod context;
-mod context_adapter;
 mod conversation_state;
 mod hooks;
 mod input_source;
